requirements.txt
protobuf
pika
bert-serving-client
sseclient-py
requests
sklearn
numpy